chore: Extract shared iOS and macOS code for ImageSource, clean ImageData cleanup
