+++ src/Uno.UI/UI/Xaml/Media/ImageSource.iOS.cs
@@ -1,40 +1,22 @@
 ﻿using System;
-using System.Diagnostics.CodeAnalysis;
 using System.IO;
+using System.Runtime.InteropServices.WindowsRuntime;
 using System.Threading;
 using System.Threading.Tasks;
-using Foundation;
-using UIKit;
-using Uno.Diagnostics.Eventing;
-using Uno.Extensions;
-using Uno.Foundation.Logging;
-using Uno.UI.Xaml.Media;
-using Windows.UI.Core;
 
 namespace Windows.UI.Xaml.Media
 {
 	public partial class ImageSource
 	{
-		private readonly bool _isOriginalSourceUIImage;
-
 		internal static readonly bool SupportsAsyncFromBundle;
 		internal static readonly bool SupportsFromBundle;
+
+		private static NSUrlSession _defaultSession;
 
 		static ImageSource()
 		{
 			SupportsAsyncFromBundle = UIDevice.CurrentDevice.CheckSystemVersion(9, 0);
 			SupportsFromBundle = UIDevice.CurrentDevice.CheckSystemVersion(8, 0);
-		}
-
-		private static NSUrlSession _defaultSession;
-		/// <summary>
-		/// The <see cref="NSUrlSession"/> to use for remote url downloads, if using the platform downloader.
-		/// By default <see cref="NSUrlSession.SharedSession"/> will be used.
-		/// </summary>
-		public static NSUrlSession DefaultSession
-		{
-			get => _defaultSession ?? NSUrlSession.SharedSession;
-			set => _defaultSession = value;
 		}
 
 		protected ImageSource()
@@ -49,49 +31,20 @@ namespace Windows.UI.Xaml.Media
 			_imageData = ImageData.FromNative(image);
 		}
 
+		/// <summary>
+		/// The <see cref="NSUrlSession"/> to use for remote url downloads, if using the platform downloader.
+		/// By default <see cref="NSUrlSession.SharedSession"/> will be used.
+		/// </summary>
+		public static NSUrlSession DefaultSession
+		{
+			get => _defaultSession ?? NSUrlSession.SharedSession;
+			set => _defaultSession = value;
+		}
+
 		internal string BundleName { get; private set; }
 		internal string BundlePath { get; private set; }
 
 		static public implicit operator ImageSource(UIImage image) => new ImageSource(image);
-
-		public bool HasSource()
-		{
-			return IsSourceReady
-				|| Stream != null
-				|| AbsoluteUri != null
-				|| FilePath.HasValueTrimmed()
-				|| _imageData.HasData
-				|| HasBundle;
-		}
-
-		/// <summary>
-		/// Determines if the current instance references a local bundle resource.
-		/// </summary>
-		public bool HasBundle => BundlePath.HasValueTrimmed() || BundleName.HasValueTrimmed();
-
-		/// <summary>
-		/// Open bundle is using either the name of the bundle (for 
-		/// android compatibility), or the path to the bundle for Windows compatibility.
-		/// </summary>
-		internal ImageData OpenBundle()
-		{
-			var image = OpenBundleFromString(BundleName) ?? OpenBundleFromString(BundlePath);
-			if (image is not null)
-			{
-				_imageData = ImageData.FromNative(image);
-			}
-			else
-			{
-				_imageData = ImageData.Empty;
-			}
-
-			if (!_imageData.HasData)
-			{
-				this.Log().ErrorFormat("Unable to locate bundle resource [{0}]", BundlePath ?? BundleName);
-			}
-
-			return _imageData;
-		}
 
 		private static UIImage OpenBundleFromString(string bundle)
 		{
@@ -103,94 +56,72 @@ namespace Windows.UI.Xaml.Media
 			return null;
 		}
 
-		/// <summary>
-		/// Indicates that this ImageSource has enough information to be opened
-		/// </summary>
-		private protected virtual bool IsSourceReady => false;
-
-		private protected virtual bool TryOpenSourceSync([NotNullWhen(true)] out ImageData image)
+		private ImageData OpenImageDataFromStream()
 		{
-			image = default;
-			return false;
-		}
-
-		private protected virtual bool TryOpenSourceAsync(CancellationToken ct, [NotNullWhen(true)] out Task<ImageData> asyncImage)
-		{
-			asyncImage = default;
-			return false;
-		}
-
-		/// <summary>
-		/// Retrieves the already loaded image, or for supported source (eg. WriteableBitmap, cf remarks),
-		/// create a native image from the data in memory.
-		/// </summary>
-		/// <remarks>
-		/// This is only intended to convert **uncompressed data already in memory**,
-		/// and should not be used to decompress a JPEG for instance, even if the already in memory.
-		/// </remarks>
-		internal bool TryOpenSync(out ImageData image)
-		{
-			if (_imageData.HasData)
+			Stream.Position = 0;
+			using var data = NSData.FromStream(Stream);
+			var nativeImage = UIImage.LoadFromData(data);
+			
+			if (nativeImage is not null)
 			{
-				image = _imageData;
-				return true;
+				return _imageData = ImageData.FromNative(nativeImage);
 			}
-
-			if (IsSourceReady && TryOpenSourceSync(out image))
+			else
 			{
-				return true;
+				return ImageData.Empty;
 			}
-
-			image = default;
-			return false;
 		}
 
-		internal async Task<ImageData> Open(CancellationToken ct)
+		private Task<ImageData> OpenImageDataFromFilePathAsync()
 		{
-			using (
-			   _trace.WriteEventActivity(
-				   TraceProvider.ImageSource_SetImageDecodeStart,
-				   TraceProvider.ImageSource_SetImageDecodeStop,
-				   new object[] { this.GetDependencyObjectId() }
-			   )
-			)
+			var uri = new Uri(FilePath);
+			var nativeImage = UIImage.FromFile(FilePath);
+
+			_imageData = nativeImage is not null ?
+				ImageData.FromNative(nativeImage) : ImageData.Empty;
+
+			return Task.FromResult(_imageData);
+		}
+
+		private async Task<ImageData> OpenImageDataFromBundleAsync()
+		{
+			if (SupportsAsyncFromBundle)
 			{
-				if (ct.IsCancellationRequested)
+				return _imageData = OpenBundle();
+			}
+			else
+			{
+				await CoreDispatcher.Main.RunAsync(
+					CoreDispatcherPriority.Normal,
+					() =>
+					{
+						_imageData = OpenBundle();
+					}
+				);
+
+				return _imageData;
+			}
+		}
+
+		private async Task<ImageData> DownloadAndOpenImageDataAsync()
+		{
+			if (Downloader == null)
+			{
+				await OpenUsingPlatformDownloader(ct);
+			}
+			else
+			{
+				var localFileUri = await Download(ct, AbsoluteUri);
+
+				if (localFileUri == null)
 				{
 					return ImageData.Empty;
 				}
 
-				if (IsSourceReady && TryOpenSourceSync(out var img))
+				if (SupportsAsyncFromBundle)
 				{
-					return _imageData = img;
-				}
-
-				if (IsSourceReady && TryOpenSourceAsync(ct, out var asyncImg))
-				{
-					return _imageData = await asyncImg;
-				}
-
-				if (Stream != null)
-				{
-					Stream.Position = 0;
-					using (var data = NSData.FromStream(Stream))
-					{
-						var nativeImage = UIImage.LoadFromData(data);
-						if (nativeImage is not null)
-						{
-							return _imageData = ImageData.FromNative(nativeImage);
-						}
-						else
-						{
-							return ImageData.Empty;
-						}
-					}
-				}
-
-				if (FilePath.HasValue())
-				{
-					var uri = new Uri(FilePath);
-					var nativeImage = UIImage.FromFile(FilePath);
+					// Since iOS9, UIImage.FromBundle is thread safe.
+					var nativeImage = UIImage.FromBundle(localFileUri.LocalPath);
 					if (nativeImage is not null)
 					{
 						return _imageData = ImageData.FromNative(nativeImage);
@@ -200,94 +131,46 @@ namespace Windows.UI.Xaml.Media
 						return _imageData = ImageData.Empty;
 					}
 				}
-
-				if (HasBundle)
-				{
-					if (SupportsAsyncFromBundle)
-					{
-						return _imageData = OpenBundle();
-					}
-					else
-					{
-						await CoreDispatcher.Main.RunAsync(
-							CoreDispatcherPriority.Normal,
-							() =>
-							{
-								_imageData = OpenBundle();
-							}
-						);
-
-						return _imageData;
-					}
-				}
-
-				if (Downloader == null)
-				{
-					await OpenUsingPlatformDownloader(ct);
-				}
 				else
 				{
-					var localFileUri = await Download(ct, AbsoluteUri);
-
-					if (localFileUri == null)
-					{
-						return ImageData.Empty;
-					}
-
-					if (SupportsAsyncFromBundle)
-					{
-						// Since iOS9, UIImage.FromBundle is thread safe.
-						var nativeImage = UIImage.FromBundle(localFileUri.LocalPath);
-						if (nativeImage is not null)
+					await CoreDispatcher.Main.RunAsync(
+						CoreDispatcherPriority.Normal,
+						() =>
 						{
-							return _imageData = ImageData.FromNative(nativeImage);
-						}
-						else
-						{
-							return _imageData = ImageData.Empty;
-						}
-					}
-					else
-					{
-						await CoreDispatcher.Main.RunAsync(
-							CoreDispatcherPriority.Normal,
-							() =>
+							if (SupportsFromBundle)
 							{
-								if (SupportsFromBundle)
+								// FromBundle calls UIImage:fromName which caches the decoded image.
+								// This is done to avoid decoding images from the byte array returned 
+								// from the cache.
+								var nativeImage = UIImage.FromBundle(localFileUri.LocalPath);
+								if (nativeImage is not null)
 								{
-									// FromBundle calls UIImage:fromName which caches the decoded image.
-									// This is done to avoid decoding images from the byte array returned 
-									// from the cache.
-									var nativeImage = UIImage.FromBundle(localFileUri.LocalPath);
-									if (nativeImage is not null)
-									{
-										_imageData = ImageData.FromNative(nativeImage);
-									}
-									else
-									{
-										_imageData = ImageData.Empty;
-									}
+									_imageData = ImageData.FromNative(nativeImage);
 								}
 								else
 								{
-									// On iOS 7, FromBundle doesn't work. We can use this method instead.
-									// Note that we must use OriginalString and not LocalPath.
-									var nativeImage = UIImage.LoadFromData(NSData.FromUrl(new NSUrl(localFileUri.OriginalString)));
-									if (nativeImage is not null)
-									{
-										_imageData = ImageData.FromNative(nativeImage);
-									}
-									else
-									{
-										_imageData = ImageData.Empty;
-									}
+									_imageData = ImageData.Empty;
 								}
-							}).AsTask(ct);
-					}
+							}
+							else
+							{
+								// On iOS 7, FromBundle doesn't work. We can use this method instead.
+								// Note that we must use OriginalString and not LocalPath.
+								var nativeImage = UIImage.LoadFromData(NSData.FromUrl(new NSUrl(localFileUri.OriginalString)));
+								if (nativeImage is not null)
+								{
+									_imageData = ImageData.FromNative(nativeImage);
+								}
+								else
+								{
+									_imageData = ImageData.Empty;
+								}
+							}
+						}).AsTask(ct);
 				}
-
-				return _imageData;
 			}
+
+			return _imageData;
 		}
 
 		/// <summary>
@@ -363,62 +246,6 @@ namespace Windows.UI.Xaml.Media
 					bool IsSuccessful(nint status) => status < 300;
 				}
 			}
-		}
-
-		/// <summary>
-		/// Similar to Dispose, but the ImageSource can still be used in the future.
-		/// </summary>
-		internal void UnloadImageData()
-		{
-			// If the original source is a UIImage, we can't dispose it because we will
-			// not be able to restore it later (from a AbsoluteUri, BundleName, file path, etc.)
-			if (!_isOriginalSourceUIImage)
-			{
-				DisposeImageData();
-			}
-		}
-
-		partial void InitFromResource(Uri uri)
-		{
-			var path = uri
-				.PathAndQuery
-				.TrimStart(new[] { '/' })
-
-				// UWP supports backward slash in path for directory separators.
-				.Replace("\\", "/");
-
-			BundlePath = path;
-
-			BundleName = uri != null
-				? Path.GetFileName(uri.AbsolutePath)
-				: null;
-		}
-
-		partial void CleanupResource()
-		{
-			BundlePath = null;
-			BundleName = null;
-		}
-
-		partial void DisposePartial()
-		{
-			DisposeImageData();
-		}
-
-		private void DisposeImageData()
-		{
-			if (_imageData.Kind == ImageDataKind.NativeImage)
-			{
-				_imageData.NativeImage?.Dispose();
-			}
-
-			_imageData = ImageData.Empty;
-		}
-
-		public override string ToString()
-		{
-			var source = Stream ?? AbsoluteUri ?? FilePath ?? (object)_imageData ?? BundlePath ?? BundleName ?? "[No source]";
-			return "ImageSource: {0}".InvariantCultureFormat(source);
 		}
 	}
 }
