+++ src/Uno.UI/UI/Xaml/Media/ImageSource.skia.cs
@@ -31,7 +31,5 @@ namespace Windows.UI.Xaml.Media
 		{
 			AbsoluteUri = null;
 		}
-
-		internal void UnloadImageData() { }
 	}
 }
